Moved the Shared Button component to themed directory
