+++ src/components/scenes/RequestScene.js
@@ -19,7 +19,6 @@ import { refreshAllFioAddresses } from '../../modules/FioAddress/action'
 import * as SETTINGS_SELECTORS from '../../modules/Settings/selectors.js'
 import type { ExchangedFlipInputAmounts } from '../../modules/UI/components/FlipInput/ExchangedFlipInput2.js'
 import { ExchangedFlipInput } from '../../modules/UI/components/FlipInput/ExchangedFlipInput2.js'
-import { ShareButtons } from '../../modules/UI/components/ShareButtons/ShareButtons.ui.js'
 import * as UI_SELECTORS from '../../modules/UI/selectors.js'
 import { THEME } from '../../theme/variables/airbitz.js'
 import { type Dispatch, type RootState } from '../../types/reduxTypes.js'
@@ -32,6 +31,7 @@ import { ButtonsModal } from '../modals/ButtonsModal.js'
 import { Airship, showError, showToast } from '../services/AirshipInstance.js'
 import { type Theme, type ThemeProps, cacheStyles, withTheme } from '../services/ThemeContext.js'
 import { EdgeText } from '../themed/EdgeText.js'
+import { ShareButtons } from '../themed/ShareButtons.js'
 import { RightChevronButton } from '../themed/ThemedButtons.js'
 
 const PUBLIC_ADDRESS_REFRESH_MS = 2000
